base m2m alterations 1 - admin only
This branch is being adapted and modified in order to serve as the
admin backend only !!

Major alterations include:

1)
There is a French version and an English version of the GoCart module
with separate db tables for each. In order to maintain data
synchronization between  the different tables and languages duplicate
db active records are implemented in certain models.

The models that force sync include :
- coupon_model
- customer_model
- gift_card_model
- order_model
- product_model
- settings_model
- SYSTEM -> session_model

Note ! Not all models are synched. Certain models maintain separation,
these include :
- banner_model
- box_model
- digital_product_model
- location_model
- routes_model
- search_model
- category_model
- category_model_mom2mom
- messages_model
- option_model
- page_model